gui interaction

+++ OreGenerator/src/de/bossascrew/generator/listener/InventoryInteractListener.java
@@ -1,5 +1,6 @@
 package de.bossascrew.generator.listener;
 
+import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -9,6 +10,8 @@ import de.bossascrew.generator.Generator;
 import de.bossascrew.generator.GeneratorObject;
 import de.bossascrew.generator.Inventories.ConfirmGUI;
 import de.bossascrew.generator.data.DataManager;
+import de.bossascrew.generator.data.Message;
+import de.bossascrew.generator.data.Permission;
 import de.tr7zw.nbtapi.NBTItem;
 
 public class InventoryInteractListener implements Listener {
@@ -17,7 +20,7 @@ public class InventoryInteractListener implements Listener {
 	@EventHandler
 	public void invClick(InventoryClickEvent e) {
 
-		if(e.getCurrentItem() == null) return;
+		if(e.getCurrentItem() == null || e.getCurrentItem().getType() == Material.AIR) return;
 		if(!(e.getView().getPlayer() instanceof Player)) return;
 		
 		NBTItem item = new NBTItem(e.getCurrentItem());
@@ -34,22 +37,34 @@ public class InventoryInteractListener implements Listener {
 			switch (action) {
 			case Generator.NBT_ACTION_VALUE_LEVELINFO:
 				if(clickedLevel > g.getLevel()) {
-					if(clickedLevel == g.getLevel()+1) {
-						p.openInventory(new ConfirmGUI(g, clickedLevel).getInv());
+					if(!p.hasPermission(Permission.LEVEL + clickedLevel)) {
+						p.sendMessage(Message.NO_PERMISSION);
 					} else {
-						deny(p);
+						if(clickedLevel == g.getLevel()+1) {
+							p.openInventory(new ConfirmGUI(g, clickedLevel).getInv());
+						} else {
+							deny(p);
+						}
 					}
 				}
 				break;
 			case Generator.NBT_ACTION_VALUE_DROP:
-				g.drop();
+				if(!p.hasPermission(Permission.DROP_GENERATOR)) {
+					p.sendMessage(Message.NO_PERMISSION);
+				} else {
+					g.drop();
+				}
 				break;
 			}
-		} else if(e.getView().getTitle() != null && !e.getView().getTitle().equals(Generator.GUI_CONFIRM_TITLE)) {
+		} else if(e.getView().getTitle() != null && e.getView().getTitle().equals(Generator.GUI_CONFIRM_TITLE)) {
 			e.setCancelled(true);
+			p.closeInventory();
+			g.open(p);
 			switch(item.getString(Generator.NBT_ACTION_KEY)) {
 			case Generator.NBT_ACTION_VALUE_CONFIRM:
+				System.out.println(item.getInteger(Generator.NBT_LEVEL_KEY));
 				if(g.tryUpgrade(item.getInteger(Generator.NBT_LEVEL_KEY))) {
+					System.out.println("Test3");
 					g.refreshGUI();
 				} else {
 					deny(p);
@@ -59,14 +74,11 @@ public class InventoryInteractListener implements Listener {
 				//TODO play sound hewwwww
 				break;
 			}				
-			p.closeInventory();
-			g.open(p);
 		}
-		
 	}
 	
 	public void deny(Player p) {
 		//TODO Villager sound
-		p.sendMessage(Generator.CANT_AFFORD_LEVEL);
+		p.sendMessage(Message.CANT_AFFORD_LEVEL);
 	}
 }
